backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
pytz==2023.3 